update hotshot to tag 0.5.24
