+++ Cargo.toml
@@ -40,7 +40,6 @@ sql-data-source = [
 	"native-tls",
 	"postgres-native-tls",
 	"refinery",
-	"serde_json",
 	"tokio",
 	"tokio-postgres",
 ]
@@ -61,12 +60,14 @@ required-features = ["sql-data-source", "testing"]
 
 [dependencies]
 anyhow = "1.0"
+ark-serialize = "0.4.2"
 async-compatibility-layer = { git = "https://github.com/EspressoSystems/async-compatibility-layer.git", tag = "1.4.1", features = [
 	"logging-utils",
 ] }
 async-std = { version = "1.9.0", features = ["unstable", "attributes"] }
 async-trait = "0.1"
 bincode = "1.3"
+bit-vec = { version = "0.6.3", features = ["serde_std"] }
 chrono = "0.4"
 clap = { version = "4.5", features = ["derive", "env"] }
 cld = "0.5"
@@ -83,6 +84,7 @@ itertools = "0.12.1"
 jf-primitives = { git = "https://github.com/EspressoSystems/jellyfish", tag = "0.4.1" }
 prometheus = "0.13"
 serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 snafu = { version = "0.7", features = ["backtraces"] }
 surf-disco = { git = "https://github.com/EspressoSystems/surf-disco.git", tag = "v0.4.6" }
 tide-disco = { git = "https://github.com/EspressoSystems/tide-disco.git", tag = "v0.4.6" }
@@ -91,8 +93,6 @@ toml = "0.8"
 tracing = "0.1"
 trait-variant = "0.1"
 typenum = "1" 
-ark-serialize = "0.4.2"
-bit-vec = { version = "0.6.3", features = ["serde_std"] }
 
 # Dependencies enabled by feature "file-system-data-source".
 atomic_store = { git = "https://github.com/EspressoSystems/atomicstore.git", tag = "0.1.3", optional = true }
@@ -102,7 +102,6 @@ include_dir = { version = "0.7", optional = true }
 native-tls = { version = "0.2", optional = true }
 postgres-native-tls = { version = "0.5", optional = true }
 refinery = { version = "0.8", features = ["tokio-postgres"], optional = true }
-serde_json = { version = "1.0", optional = true }
 tokio = { version = "1.36", optional = true }
 tokio-postgres = { version = "0.7", optional = true, default-features = false, features = [ # disabling the default features removes dependence on the tokio runtime
 	"with-serde_json-1",
