+++ examples/simple-server.rs
@@ -146,9 +146,9 @@ async fn init_consensus(
             state_ver_key: state_key_pair.ver_key(),
         })
         .collect::<Vec<_>>();
+
+    let num_nodes_with_stake = NonZeroUsize::new(pub_keys.len()).unwrap();
     let config = HotShotConfig {
-        num_nodes_with_stake: NonZeroUsize::new(pub_keys.len()).unwrap(),
-        num_nodes_without_stake: 0,
         known_nodes_with_stake: known_nodes_with_stake.clone(),
         known_nodes_without_stake: vec![],
         start_delay: 0,
@@ -163,8 +163,11 @@ async fn init_consensus(
         execution_type: ExecutionType::Continuous,
         election_config: None,
         da_staked_committee_size: pub_keys.len(),
-        da_non_staked_committee_size: 0,
         my_own_validator_config: Default::default(),
+        num_nodes_with_stake,
+        num_nodes_without_stake: 0,
+        known_nodes_without_stake: Vec::new(),
+        da_non_staked_committee_size: 0,
     };
     join_all(priv_keys.into_iter().zip(data_sources).enumerate().map(
         |(node_id, (priv_key, data_source))| {
@@ -185,10 +188,8 @@ async fn init_consensus(
                     state_key_pair: state_key_pairs[node_id].clone(),
                 };
 
-                let election_config = MockMembership::default_election_config(
-                    config.num_nodes_with_stake.get() as u64,
-                    0,
-                );
+                let election_config =
+                    MockMembership::default_election_config(pub_keys.len() as u64, 0);
                 let membership = MockMembership::create_election(
                     known_nodes_with_stake.clone(),
                     election_config,
