+++ src/data_source/update.rs
@@ -74,13 +74,20 @@ where
             let qcs = once((**qc).clone())
                 // ...and each leaf in the chain justifies the subsequent leaf (its parent) through
                 // `leaf.justify_qc`.
-                .chain(leaf_chain.iter().map(|leaf| leaf.leaf.get_justify_qc()))
+                .chain(
+                    leaf_chain
+                        .iter()
+                        .map(|leaf_info| leaf_info.leaf.get_justify_qc()),
+                )
                 // Put the QCs in chronological order.
                 .rev()
                 // The oldest QC is the `justify_qc` of the oldest leaf, which does not justify any
                 // leaf in the new chain, so we don't need it.
                 .skip(1);
             for (qc, leaf_info) in qcs.zip(leaf_chain.iter().rev()) {
+                let leaf = leaf_info.leaf.clone();
+                let vid = leaf_info.vid.clone();
+
                 // `LeafQueryData::new` only fails if `qc` does not reference `leaf`. We have just
                 // gotten `leaf` and `qc` directly from a consensus `Decide` event, so they are
                 // guaranteed to correspond, and this should never panic.
